requirements.txt

alembic==1.10.4
amqp==5.1.1
async-timeout==4.0.2
Babel==2.12.1
billiard==3.6.4.0
blinker==1.6.2
cachelib==0.9.0
celery==5.2.7
cffi==1.15.1
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
colorama==0.4.6
cryptography==40.0.2
dnspython==2.3.0
dominate==2.7.0
email-validator==2.0.0.post2
Flask==2.2.1
Flask-BabelEx==0.9.4
Flask-Bootstrap==3.3.7.1
Flask-CacheControl==0.3.0
Flask-Caching==2.0.2
Flask-JSON==0.3.5
Flask-Login==0.6.2
Flask-Mail==0.9.1
Flask-Migrate==4.0.4
Flask-Moment==1.0.5
flask-mongoengine==1.0.0
Flask-Principal==0.4.0
Flask-Security==3.0.0
Flask-SQLAlchemy==3.0.3
Flask-WTF==1.1.1
greenlet==2.0.2
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
kombu==5.2.4
Mako==1.2.4
MarkupSafe==2.1.2
mongoengine==0.27.0
packaging==23.1
passlib==1.7.4
prompt-toolkit==3.0.38
psycopg2==2.9.6
pycparser==2.21
pymongo==4.3.3
PyMySQL==1.0.3
pypi==2.1
python-dotenv==1.0.0
pytz==2023.3
redis==4.5.5
six==1.16.0
speaklater==1.3
SQLAlchemy==2.0.12
typing_extensions==4.5.0
vine==5.0.0
visitor==0.1.3
wcwidth==0.2.6
Werkzeug==2.3.2
WTForms==3.0.1
